add novas funcionalidades e correções

+++ app/src/main/java/quixada/ufc/br/kisan/adapter/OnCustomClickListener.java
@@ -5,5 +5,5 @@ import android.view.View;
 
 public interface OnCustomClickListener {
 
-    public void OnCustomClick(View view, int position);
+    public void OnCustomClick(View view, Long position);
 }
